image_generator/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astroid==3.0.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dill==0.3.7
distlib==0.3.6
Django==5.0
django-cors-headers==4.3.1
django-decouple==2.1
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
environs==10.3.0
fastapi==0.110.0
fastapi_cors==0.0.6
filelock==3.12.2
Flask==3.0.0
Flask-WTF==1.2.1
git-filter-repo==2.38.0
h11==0.14.0
httptools==0.6.1
idna==3.6
install==1.3.5
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.3
marshmallow==3.21.0
mccabe==0.7.0
numpy==1.24.3
packaging==23.2
Pillow==10.1.0
platformdirs==3.8.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.36.3
tomlkit==0.12.3
typing_extensions==4.10.0
tzdata==2023.3
unicorn==2.0.1.post1
uvicorn==0.27.1
vine==5.1.0
virtualenv==20.23.1
virtualenvwrapper-win==1.2.7
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.1
